Пофиксил баг при котором во время состояния busy при работе с БД в интерфейсе отображалось неадекватная информация

+++ Qt/myConvolution/backend.cpp
@@ -13,7 +13,7 @@ QString Backend::lastError()
 void Backend::setDbStatus(int const &dbStatus) {
     if (_dbStatus == dbStatus)
         return;
-    _dbStatus = static_cast<int>(dbStatus);
+    _dbStatus = dbStatus;
 
     emit dbStatusChanged(dbStatus);
 }
